refactor: split and simplify BPF data plane code
Split the monolithic xdp_prog.c (867 lines) into focused modules
using the #include pattern, and consolidate duplicated logic:

File structure:
- xdp_prog.c: slim orchestrator (~48 lines, includes only)
- endpoint/tailcall_endpoint.c: 16 endpoint SEC programs
- headend/tailcall_headend.c: 7 headend SEC programs
- dispatch/srv6_dispatch.c: 4 SRv6 dispatchers
- dispatch/l2_headend.c: L2 headend pipeline
- xdp_main.c: process_l3 + vinbero_main entry point
- l2vpn/tc_prog.c: TC ingress entry point

Code simplification:
- Merge H.Encaps + H.Encaps.Red into unified do_h_encaps_impl(bool reduced),
  delete srv6_encaps_red.h
- Merge TC BUM encap functions into tc_do_single_pe_encap_impl(bool reduced)
- Extract fib_result_to_xdp_action() helper, replacing 8 duplicated switches
- Add DEFINE_ENDPOINT_LOCALSID{,_AUX} macros for Pattern A endpoints
- Unify HEADEND_V4_BODY/V6_BODY into HEADEND_BODY(fn, hdr_type)
