--- src/headend/srv6_encaps_red.h
@@ -1,181 +1,0 @@
-#ifndef SRV6_ENCAPS_RED_H
-#define SRV6_ENCAPS_RED_H
-
-#include <linux/types.h>
-#include <linux/if_ether.h>
-#include <linux/ipv6.h>
-#include <linux/in.h>
-#include <bpf/bpf_helpers.h>
-#include <bpf/bpf_endian.h>
-
-#include "core/xdp_prog.h"
-#include "core/srv6.h"
-#include "headend/srv6_headend_utils.h"
-#include <linux/ip.h>
-#include "core/srv6_fib.h"
-
-// H.Encaps.Red single-segment: no SRH, just outer IPv6 with nexthdr = inner_proto
-static __always_inline int do_h_encaps_red_1seg(
-    struct xdp_md *ctx,
-    struct ethhdr *saved_eth,
-    struct headend_entry *entry,
-    __u8 inner_proto,
-    __u16 inner_total_len,
-    __u16 l3_offset)
-{
-    int new_headers_len = sizeof(struct ipv6hdr);
-    int vlan_len = l3_offset - ETH_HLEN;
-
-    if (bpf_xdp_adjust_head(ctx, -(new_headers_len - vlan_len)))
-        return XDP_DROP;
-
-    void *data = (void *)(long)ctx->data;
-    void *data_end = (void *)(long)ctx->data_end;
-
-    struct ethhdr *new_eth = data;
-    CHECK_BOUND(new_eth, data_end, sizeof(*new_eth));
-
-    struct ipv6hdr *outer_ip6h = (struct ipv6hdr *)(new_eth + 1);
-    CHECK_BOUND(outer_ip6h, data_end, sizeof(*outer_ip6h));
-
-    build_outer_ipv6(outer_ip6h, inner_proto, inner_total_len,
-                     entry->src_addr, &entry->segments[0]);
-
-    __builtin_memcpy(new_eth, saved_eth, sizeof(struct ethhdr));
-    new_eth->h_proto = bpf_htons(ETH_P_IPV6);
-
-    __u32 ifindex;
-    int fib_result = srv6_fib_lookup_and_update(ctx, outer_ip6h, new_eth, &ifindex, ctx->ingress_ifindex);
-
-    switch (fib_result) {
-    case FIB_RESULT_REDIRECT:
-        return bpf_redirect(ifindex, 0);
-    case FIB_RESULT_DROP:
-        return XDP_DROP;
-    default:
-        return XDP_PASS;
-    }
-}
-
-// H.Encaps.Red multi-segment: reduced SRH with N-1 entries (S1 omitted)
-static __always_inline int do_h_encaps_red_multi(
-    struct xdp_md *ctx,
-    struct ethhdr *saved_eth,
-    struct headend_entry *entry,
-    __u8 inner_proto,
-    __u16 inner_total_len,
-    __u16 l3_offset)
-{
-    if (entry->num_segments < 2 || entry->num_segments > MAX_SEGMENTS)
-        return XDP_DROP;
-
-    __u8 reduced_count = entry->num_segments - 1;
-    if (reduced_count < 1 || reduced_count > 9)
-        return XDP_DROP;
-    int srh_len = 8 + (16 * (int)reduced_count);
-    int new_headers_len = (int)sizeof(struct ipv6hdr) + srh_len;
-    int vlan_len = l3_offset - ETH_HLEN;
-
-    if (bpf_xdp_adjust_head(ctx, -(new_headers_len - vlan_len)))
-        return XDP_DROP;
-
-    void *data = (void *)(long)ctx->data;
-    void *data_end = (void *)(long)ctx->data_end;
-
-    struct ethhdr *new_eth = data;
-    CHECK_BOUND(new_eth, data_end, sizeof(*new_eth));
-
-    struct ipv6hdr *outer_ip6h = (struct ipv6hdr *)(new_eth + 1);
-    CHECK_BOUND(outer_ip6h, data_end, sizeof(*outer_ip6h));
-
-    struct ipv6_sr_hdr *srh = (struct ipv6_sr_hdr *)(outer_ip6h + 1);
-    CHECK_BOUND(srh, data_end, 8);
-    CHECK_BOUND(srh, data_end, srh_len);
-
-    build_outer_ipv6(outer_ip6h, IPPROTO_ROUTING, srh_len + inner_total_len,
-                     entry->src_addr, &entry->segments[0]);
-
-    srh->nexthdr = inner_proto;
-    srh->hdrlen = (srh_len >> 3) - 1;
-    srh->type = IPV6_SRCRT_TYPE_4;
-    srh->segments_left = reduced_count;
-    srh->first_segment = reduced_count - 1;
-    srh->flags = 0;
-    srh->tag = 0;
-
-    void *srh_segments = (void *)srh + 8;
-    if (copy_segments_to_srh_reduced(srh_segments, data_end, entry->segments, entry->num_segments) != 0)
-        return XDP_DROP;
-
-    __builtin_memcpy(new_eth, saved_eth, sizeof(struct ethhdr));
-    new_eth->h_proto = bpf_htons(ETH_P_IPV6);
-
-    __u32 ifindex;
-    int fib_result = srv6_fib_lookup_and_update(ctx, outer_ip6h, new_eth, &ifindex, ctx->ingress_ifindex);
-
-    switch (fib_result) {
-    case FIB_RESULT_REDIRECT:
-        return bpf_redirect(ifindex, 0);
-    case FIB_RESULT_DROP:
-        return XDP_DROP;
-    default:
-        return XDP_PASS;
-    }
-}
-
-// H.Encaps.Red core dispatcher (RFC 8986 Section 5.1.1)
-static __always_inline int do_h_encaps_red_core(
-    struct xdp_md *ctx,
-    struct ethhdr *saved_eth,
-    struct headend_entry *entry,
-    __u8 inner_proto,
-    __u16 inner_total_len,
-    __u16 l3_offset)
-{
-    if (entry->num_segments == 1)
-        return do_h_encaps_red_1seg(ctx, saved_eth, entry, inner_proto, inner_total_len, l3_offset);
-
-    return do_h_encaps_red_multi(ctx, saved_eth, entry, inner_proto, inner_total_len, l3_offset);
-}
-
-// H.Encaps.Red for IPv4 (RFC 8986 Section 5.1.1)
-static __always_inline int do_h_encaps_red_v4(
-    struct xdp_md *ctx,
-    struct ethhdr *eth,
-    struct iphdr *iph,
-    struct headend_entry *entry,
-    __u16 l3_offset)
-{
-    if (entry->num_segments < 1 || entry->num_segments > MAX_SEGMENTS) {
-        DEBUG_PRINT("H.Encaps.Red.v4: Invalid segment count %d\n", entry->num_segments);
-        return XDP_DROP;
-    }
-
-    struct ethhdr saved_eth;
-    __builtin_memcpy(&saved_eth, eth, sizeof(struct ethhdr));
-    __u16 inner_total_len = bpf_ntohs(iph->tot_len);
-
-    return do_h_encaps_red_core(ctx, &saved_eth, entry, IPPROTO_IPIP, inner_total_len, l3_offset);
-}
-
-// H.Encaps.Red for IPv6 (RFC 8986 Section 5.1.1)
-static __always_inline int do_h_encaps_red_v6(
-    struct xdp_md *ctx,
-    struct ethhdr *eth,
-    struct ipv6hdr *inner_ip6h,
-    struct headend_entry *entry,
-    __u16 l3_offset)
-{
-    if (entry->num_segments < 1 || entry->num_segments > MAX_SEGMENTS) {
-        DEBUG_PRINT("H.Encaps.Red.v6: Invalid segment count %d\n", entry->num_segments);
-        return XDP_DROP;
-    }
-
-    struct ethhdr saved_eth;
-    __builtin_memcpy(&saved_eth, eth, sizeof(struct ethhdr));
-    __u16 inner_total_len = 40 + bpf_ntohs(inner_ip6h->payload_len);
-
-    return do_h_encaps_red_core(ctx, &saved_eth, entry, IPPROTO_IPV6, inner_total_len, l3_offset);
-}
-
-#endif // SRV6_ENCAPS_RED_H
